requirements.txt
gspread
spacy
GitPython